refactor: Phase 14 — expand astra-plan with plan lifecycle + metrics + performance
- Move plan/mod.rs (1356 LOC) to astra-plan as plan.rs
- Move plan/metrics.rs (302 LOC) and plan/performance.rs (449 LOC)
- Fix module paths (super::decompose → crate::decompose, metrics:: → crate::metrics::)
- astra-plan: 182 → 227 tests (+45)
- Runtime: 4,018 → 3,973 tests (transferred to crates)
