+++ rust/crates/astra-plan/src/lib.rs
@@ -1,5 +1,9 @@
 //! Astra Plan — goal decomposition engine for breaking complex tasks into subtasks.
 
 pub mod decompose;
+pub mod metrics;
+pub mod performance;
+pub mod plan;
 
 pub use decompose::*;
+pub use plan::*;
